fix: R28 全局边缘用例修复 — 31个文件136行
P1级别修复:
- DataLogger seek() 4处无返回值检查，文件损坏风险
- OtaManager HEX转BIN后显示临时文件名而非原始文件名
- QuickCommandBar 编辑/添加/清空后未持久化到QSettings
- TcpConnection/UdpConnection sender() qobject_cast未判空
- ProtocolView selectionModel() 3处未判空
- MainWindowSignalConnect 重连倒计时整数除法丢失小数

P2级别修复:
- QDataStream 3处缺少setVersion(Qt_6_8)版本固定
- DataExporter/ProtocolView CSV BOM写入和导出写入未检查返回值
- XModem/YModem/ZModem switch缺少default分支
- IntelHexParser switch缺少default分支
- ConnectionControllerHandlers/FrameParser等null检查增强
- ChannelConfig/FrameDefinition/ChartWidget等bounds检查增强

评分: 78 + 1 = 79
变更: 31 files, 109 insertions, 27 deletions

+++ src/ota/protocols/XModemTransferHandlers.cpp
@@ -61,7 +61,7 @@ void XModemTransfer::handleStateSendingBlock(char ch, int& readIdx)
         // 更新速率统计
         updateTransferStats();
 
-        int percent = static_cast<int>((m_bytesSent * 100) / m_data.size());
+        int percent = static_cast<int>((static_cast<qint64>(m_bytesSent) * 100) / m_data.size());
         emit progress(percent, m_bytesSent, m_data.size());
 
         if (m_bytesSent >= m_data.size()) {
